cviceni/cv_5/producent/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Jan Kolomazník", email = "jan.kolomaznik@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "pika>=1.3.2",
    "python-logstash>=0.4.8",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "uvicorn src.main:app --reload"

